Don't clobber existing key/csr on openssl failure
openssl doesn't seem to reliably set a return code on failure. Previously
we deleted the target key/csr file before running so we could detect their
re-creation. But this clobbers existing files following failure. Change
the logic to output key/csr to temporary files (which we are free to delete
in advance) and then rename them afterwards.

The only obvious downside to this is that openssl annoyingly says "writing
new private key to 'l.key.new'" which we'd rather keep quiet about...